Update README.md
Well, I was challenged with the following question:
How if one invests into QQQ in year 2000, would one make money by end of year 2010?

This is a good question as QQQ price was high in year 2000 and low in year 2010.
The result when investing USD 2 million within 1 year into QQQ ETF is added. The result is from a revised python file "DollarCostAveraging_QQQ_2000_2010.py".