Update README.md
A Python program titled DollarCostAveraging_2000_2010_Window_12m_to_48m.py has been added to this repository for this study.

The results are added. We can observe that for investment windows of 36 months and 48 months, there are no negative returns, indicating that longer DCA periods may help mitigate the risk of poor market timing.